requirements.txt
APScheduler==3.6.3
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
codext==1.5.5
colorama==0.4.4
crayons==0.3.1
cryptography==3.1.1
decorator==4.4.2
idna==2.10
markdown2==2.3.9
piexif==1.1.3
Pillow==6.2.2
pycodestyle==2.6.0
pycparser==2.20
python-telegram-bot==13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
stegano==0.9.8
toml==0.10.1
tornado==6.0.4
tzlocal==2.1
urllib3==1.25.10
